Files from CSS/422: Software Architecture
CSS/422 covered a variety of different software architecture styles and has myself and my learning team design the architecture for an online store, which involved the creation of a variety of different diagrams such as; Use Case diagrams, Client Server diagrams, Flow diagrams, and Architectural Pattern diagrams. The assignment culminated in a final Software Reference Guide that could guide a development team through the macro steps of the systems design, and included input about client requirements, recommendations for team size and budget, and described the decision process behind choosing the architectural style for the final product.