Files from MBL/400: Mobile Fundamentals
MBL/400 focused on reviewing and improving upon existing mobile apps, and creating a concept and wireframes for a new mobile app with my learning team. The app review process was an individual assignment that involved reviewing the quality of existing apps using industry standards, creating wireframes for any changes or improvements, and discussing my experiences with apps that I am familiar with and making recommendations for any changes that I think could improve the app. My learning team decided to create a mobile app for purchasing and reading Comic Books that we named 'The Comic Hoard', and we followed the theme of a hoard of Comic Books to create an app that encouraged users to take pride in the Comics that they own both in digital and print form. This involved creating a unique idea for an app with a specific theme, and creating Wire-Frames for the development to use as an example of how the final product should look.