Files from DBM/380: Database Concepts
DBM/300 focused on designing and creating databases for business systems that the students are familiar with. For my final assignment I chose to create a new database for Starbucks' inventory management system with some improvements that focus on the user's experience, and generating reports that contain useful information for Store Managers and Shift Supervisors about product waste, sell quantity vs. pull quantity, and recommended product pars for each food item that Starbucks sells at the store that I supervise. This involved creating and populating a database of a variety of different Starbucks food products, creating SQL queries and reports to generate and provide useful information, and Entity Relationship diagrams (ERDs) that describe the relationships between different tables in the server that focused on quality control and avoiding data entry error.