Files from CMGT/410: Project Planning and Implementation
CMGT/410 focused on designing and implementing software development projects using Microsoft Project, and by creating a variety of documentation such as a Business Requirements document, Project Proposals and Outlines, and Work Breakdown Structures that included Resource Sheets, Gantt Charts, and Team Planners. This is where my interest really began to peak during my academic pursuits, because this class focused on Project Management and Design. 
The individual assignment for this class involved creating an entire project plan for the implementation and training of a new HR system for a mid level business. It involved creating multiple Work Breakdown Structures, Resource Sheets, Team Planners, and even comparisons between different methods of development for the same project proposal. 
The team assignment focused on creating a Project Proposal, Business Requirements Document, and Work Breakdown Structure for a smart watch company's new mobile app, where we made recommendations for the components and development plan/style that should be included in the apps final product.