Files from CMGT/445: Application Implementation
CMGT 445 focused on the specifics of Application Implementation, and involved the creation of a complete Implementation Plan for a new Business Resource Planning System for an Enterprise level company. This assignment involved the identification of Key Stake Holders for the new system, a communications plan for communicating between the Key Stake-Holders and the development team, creation of project milestones and deliverables, a Change Control Process for any changes that need to be made during the software's development life cycle, and a complete Work Breakdown Structure for the project that set specific dates for each Milestone and Deliverable for the project. It also included budgeting for the training of all of the companies managers that was broken down in a resource sheet to give a basic estimate for the price of the development and training for the new system. 
This class really honed in on the various standards and methods that need to be set before a project begins in order for it to run smoothly. It included information about how changes should be handled, specific dates for training branch managers and key stakeholders, and even specific resource breakdowns for budgeting on things like Hotel rooms for trainees, food for the training days, pay for the employees on an Over-Time included weekend of training, and housing and event hosting costs for two separate  training events that allowed flexibility for all Managers that needed to be trained for this new system.